add alert msg before delete

+++ frontend/travello-frontend/src/admin/showDetails/activity/ShowActivityAgency.js
@@ -27,8 +27,12 @@ export default function ShowActivityAgency() {
   }, []);
 
   const handleDelete = async (id) => {
-    await AdminAPI.deleteFromDB(type, id);
-    fetchData().then(() => {});
+    let isExecuted = window.confirm("Are you sure to execute this action?");
+    console.log(isExecuted);
+    if (isExecuted) {
+      await AdminAPI.deleteFromDB(type, id);
+      fetchData().then(() => {});
+    }
   };
 
   const handleAddModalShow = () => {
